requirements.txt
paho-mqtt
yaml
requests